LadarIbeoLux: Improve detection of magic word in Lux stream data
Every new data block sent by the Lux begins with the magic word. The old
detection method was inefficient in that it read 4 bytes of the network
stream and repeated that up to 1000 times until the magic word was found.
Improve the detection to byte-wise comparison so that the next occurence
of the magic word is found at all times.