Delete all files generated by autogen.sh.
autogen.sh uses the autoconf version found on the host PC. As different versions
of autoconf generate differing contents of files under their control it was a pain
to update those files in the repository. The core RACK developer decided to remove
the autoconf generated files, mostly GNUmakefile.in in nearly all directories, from
the repo.

Additionally, a bunch of .gitignore files were added for not pushing those files in
again.